Testing develop branch

+++ src/main/java/dacd/riveromonzon/practice1/model/Location.java
@@ -2,10 +2,10 @@ package dacd.riveromonzon.practice1.model;
 
 public class Location {
 	private String name;
-	private double latitude;
-	private double longitude;
+	private Float latitude;
+	private Float longitude;
 
-	public Location(String name, double latitude, double longitude) {
+	public Location(String name, Float latitude, Float longitude) {
 		this.name = name;
 		this.latitude = latitude;
 		this.longitude = longitude;
